server/requirements.txt
asgiref==3.11.1
channels>=4.2
channels-redis>=4.2
redis>=5.0
daphne>=4.1
Django==6.0.4
django-cors-headers==4.9.0
django-jazzmin==3.0.4
django-rest-framework==0.1.0
djangorestframework==3.17.1
pillow==12.2.0
sqlparse==0.5.5
tzdata==2026.1
